backend/requirements.txt
atomicwrites==1.4.1
Django==3.2
django-filter==22.1
djangorestframework==3.12.4
flake8==5.0.4
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
requests==2.26.0
sqlparse==0.4.3
typing==3.7.4.3
typing_extensions==4.4.0
urllib3==1.26.14
djoser==2.1.0

python-dotenv
psycopg2-binary==2.9.3
